feat(file_name): 2 tests instead of one
